+++ test/type_inference/init.lua
@@ -992,3 +992,8 @@ local mt
 ---@type number
 function mt:f(<?x?>) end
 ]]
+
+TEST 'Test' [[
+---@class Test
+_G.<?Test?> = {}
+]]