+++ changelog.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## 2.4.10
+`2021-11-23`
 * `FIX` [#790](https://github.com/sumneko/lua-language-server/issues/790)
 * `FIX` [#798](https://github.com/sumneko/lua-language-server/issues/798)
 * `FIX` [#804](https://github.com/sumneko/lua-language-server/issues/804)