+++ changelog.md
@@ -29,6 +29,7 @@
 * `CHG` after using `Lua.runtime.nonstandardSymbol` to treat `//` as a comment, `//` is no longer parsed as an operator
 
 ## 2.4.10
+`2021-11-23`
 * `FIX` [#790](https://github.com/sumneko/lua-language-server/issues/790)
 * `FIX` [#798](https://github.com/sumneko/lua-language-server/issues/798)
 * `FIX` [#804](https://github.com/sumneko/lua-language-server/issues/804)